Rename CN project1.pptx to semester4/Computer networks/ project1.pptx
This presentation is on HTTPS and web services. HTTPS (HyperText Transfer Protocol Secure) is a secure version of HTTP that encrypts data between a user's browser and a website to ensure privacy and security. Web services are tools that allow different applications to communicate and exchange data over the internet using standard protocols like HTTP, XML, or JSON.