Add files via upload
This project 3RP is based on the principles reduce, reuse and recycle. The main idea is to collect E waste, which is increasing through time, and extract the metal ores (like alluminium, copper ,etc). The metals waste such as computer parts,electronic gadgets,etc.) are reduced by reusing them to recycle the metal ores in them. The metal ores can be sold at relatively less cost.